Prototype truncated breadcrumb
